refactor: split page file
